CC2420 and packet radio API cleanup
git-svn-id: https://mspsim.svn.sourceforge.net/svnroot/mspsim/mspsim@265 23d1a52b-0c3c-0410-b72d-8f29ab48fe35

+++ se/sics/mspsim/chip/PacketListener.java
@@ -40,5 +40,5 @@ package se.sics.mspsim.chip;
 
 public interface PacketListener {
   public void transmissionStarted();
-  public void transmissionEnded(int[] receivedData);
+  public void transmissionEnded(byte[] receivedData);
 }
